generalise support for usb audio delete old temp synth files on pi

+++ jack_script.sh
@@ -8,6 +8,8 @@ if [ -f /etc/rpi-issue ]
         raspberry=1
         export SC_JACK_DEFAULT_INPUTS="system"
         export SC_JACK_DEFAULT_OUTPUTS="system"
+        killall jackd
+        sleep 5
     else
         raspberry=0
 fi
@@ -50,19 +52,34 @@ if aplay -l | grep -qi ultra
             jackd  -d alsa -d hw:Ultra -r 44100 &
    fi
   else
-	#start jack with default hardware
-	#jackd  -d alsa -d hw:0 -r 44100 &
-    if [ $raspberry -ne 0 ]
+    # are there any other USB devices?
+
+    if aplay -l | grep -qi usb
         then
-            #amixer cset numid=3 1
-            #sleep 1
-            killall jackd
-            sleep 5
-            ( jackd -T -p 32 -d alsa -d hw:0,0 -r 44100 -p 1024 -n3  -s -i 0 -P || ( killall jackd ; sleep 10 ; jackd -T -p 32 -d alsa -d hw:0,0 -r 44100 -p 1024 -n3  -s -i 0 -P || sudo shutdown -r now )) &
-    fi
-    if [ $raspberry -eq 0  ]
-        then
-        	jackd -p32 -dalsa -dhw:0,0 -p1024 -n3 -s &
+            # get the device name
+            device=`aplay -l |grep -i USB | sed -n 1p | awk '{print $3}'`
+            if [ $raspberry -ne 0 ]
+                then
+                    ( jackd  -d alsa -d hw:$device -r 44100 -i 0 -P || ( killall jackd ; sleep 10 ; jackd  -d alsa -d hw:$device -r 44100 -i 0 -P || sudo shutdown -r now ) ) &
+                else
+                    #normal computer
+                    jackd  -d alsa -d hw:$device -r 44100 -P &
+            fi
+        else
+
+	        #start jack with default hardware
+	        #jackd  -d alsa -d hw:0 -r 44100 &
+            if [ $raspberry -ne 0 ]
+                then
+                    #amixer cset numid=3 1
+                    #sleep 1
+
+                    ( jackd -T -p 32 -d alsa -d hw:0,0 -r 44100 -p 1024 -n3  -s -i 0 -P || ( killall jackd ; sleep 10 ; jackd -T -p 32 -d alsa -d hw:0,0 -r 44100 -p 1024 -n3  -s -i 0 -P || sudo shutdown -r now )) &
+            fi
+            if [ $raspberry -eq 0  ]
+                then
+                	jackd -p32 -dalsa -dhw:0,0 -p1024 -n3 -s &
+            fi
     fi
 fi
